moving stuff to helper and added irradiance tags
