+++ src/main/java/org/patryk3211/powergrid/electricity/solarpanel/SolarPanelBlockEntity.java
@@ -5,9 +5,7 @@ import dev.ryanhcode.sable.companion.SableCompanion;
 import dev.ryanhcode.sable.sublevel.SubLevel;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.SectionPos;
-import net.minecraft.tags.BlockTags;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.chunk.ChunkAccess;
@@ -20,7 +18,7 @@ import org.patryk3211.powergrid.electricity.base.Rotation4ElectricBlock;
 import org.patryk3211.powergrid.electricity.base.ThermalBehaviour;
 import org.patryk3211.powergrid.electricity.sim.node.VoltageSourceCoupling;
 
-import static org.patryk3211.powergrid.electricity.solarpanel.SolarPanelBearingBlockEntity.DDA;
+import static org.patryk3211.powergrid.electricity.solarpanel.SolarHelper.*;
 
 public abstract class SolarPanelBlockEntity extends ElectricBlockEntity {
     protected VoltageSourceCoupling sourceCoupling;
@@ -70,14 +68,7 @@ public abstract class SolarPanelBlockEntity extends ElectricBlockEntity {
                 AMBIENT_TEMP = 22f;
             firstTick = false;
         }
-
-        if (world.isRaining() && world.isThundering()) {
-            cloudCover = .925f;
-        } else if (world.isRaining()) {
-            cloudCover = .85f;
-        } else {
-            cloudCover = 0;
-        }
+        cloudCover = getWeather(world);
 
         getPlacedBlockRotation();
         final SubLevel subLevel = Sable.HELPER.getContaining(this);
@@ -87,9 +78,9 @@ public abstract class SolarPanelBlockEntity extends ElectricBlockEntity {
         }
 
         var irradiance = getIrradiance(getAM(world), cloudCover, pos.getY(), world);
-        var cellTemp = getCellTemp(irradiance);
+        var cellTemp = getCellTemp(irradiance, AMBIENT_TEMP);
         var Vt = 8.617e-5 * (cellTemp + 273.15);
-        double[] adjusted = getTempAdjusted(irradiance, cellTemp, Vt);
+        double[] adjusted = getTempAdjusted(irradiance, cellTemp, Vt, STRINGS_IN_PARALLEL);
         double cellCurrent = adjusted[0];
         double Voc_t = adjusted[1];
         double Voc_panel = Voc_t * totalCells;
@@ -119,18 +110,6 @@ public abstract class SolarPanelBlockEntity extends ElectricBlockEntity {
         super.electricalTick();
     }
 
-    public static double[] getTempAdjusted(double irradiance, double cellTemp, double Vt) {
-        var Isc_T = SHORT_CURRENT * STRINGS_IN_PARALLEL * (irradiance / 1000) * (1 + ALPHAISC * (cellTemp - 25));
-        if (Isc_T <= 0) return new double[]{0, 0};
-        var Voc_base = IDEALITY * Vt * Math.log(Isc_T / I_O + 1);
-        var Voc_T = Voc_base + BETAVOC * (cellTemp - 25);
-        return new double[]{Isc_T, Voc_T};
-    }
-
-    public double getCellTemp(double Irradiance){
-        return AMBIENT_TEMP + (NOCT - 20) * (Irradiance / 800);
-    }
-
     public double getIrradiance(double AM, double cloudCover, int YPos, Level world) {
         if (AM == Double.POSITIVE_INFINITY) return 0;
         var transmisttance = 1 - cloudCover;
@@ -147,26 +126,11 @@ public abstract class SolarPanelBlockEntity extends ElectricBlockEntity {
         }
 
         double cosIncidence = Math.max(0, sunDir.dot(panelNormal));
-
         cosIncidence = Math.max(0, cosIncidence);
-
         double diffuseLight = 0.1 * irradiance * (1 + cloudCover) * ((1 + panelNormal.y()) / 2);
         double reflected = 0.15 * irradiance * ((1 - panelNormal.y()) / 2.0);
 
         return (irradiance * sunVisablity) * transmisttance * cosIncidence + diffuseLight +  reflected;
-    }
-
-    public double getAM(Level world){
-        var sunAngle = Math.max(0, Math.cos(world.getSunAngle(0)));
-        if (sunAngle <= 0) {
-            return Double.POSITIVE_INFINITY;
-        }
-        double elevationRad = Math.asin(sunAngle);
-        double elevationDeg = Math.toDegrees(elevationRad);
-
-        // Kasten-Young formula
-        var result = 1.0 / (sunAngle + 0.50572 * Math.pow(elevationDeg + 6.07995, -1.6364));
-        return Math.max(1, result);
     }
 
     public float sunRaycast(Level world){
@@ -176,7 +140,6 @@ public abstract class SolarPanelBlockEntity extends ElectricBlockEntity {
         var blockPos = BlockPos.containing(d.x, d.y, d.z);
         int castLength = 0;
         ChunkAccess chunk;
-
         double sunAngle = world.getSunAngle(0);
         double sunX = -Math.sin(sunAngle);
         double sunY = Math.cos(sunAngle);
@@ -207,23 +170,18 @@ public abstract class SolarPanelBlockEntity extends ElectricBlockEntity {
                     break;
                 }
             }
-            if (blockState.is(ModdedTags.Block.GLASS_BLOCK.tag) || blockState.is(ModdedTags.Block.GLASS_PANE.tag)) {
-                returnValue *= .8f;
+            if (blockState.is(ModdedTags.Block.SOLAR_QUARTER_LIGHT.tag)) {
+                returnValue *= .25f;
                 continue;
             }
-            if (blockState.is(Blocks.WATER)) {
+            if (blockState.is(ModdedTags.Block.SOLAR_HALF_LIGHT.tag)) {
                 returnValue *= .5f;
                 continue;
             }
-            if (blockState.is(BlockTags.LEAVES)) {
-                returnValue *= .2f;
+            if (blockState.is(ModdedTags.Block.SOLAR_3QUARTER_LIGHT.tag)) {
+                returnValue *= .75f;
                 continue;
             }
-            if (blockState.is(Blocks.IRON_BARS)) {
-                returnValue *= 9f / 16;
-                continue;
-            }
-
             returnValue = 0;
             break;
         }
